refactor LanguageInterceptor and update use cases
