+++ domain/usecase/src/main/kotlin/usecase/tvshow/GetTVShowVideos.kt
@@ -1,14 +1,15 @@
 package usecase.tvshow
 
-import com.berlin.entity.Video
 import repository.TVShowDetailsRepository
 import javax.inject.Inject
 class GetTVShowVideos @Inject constructor(
     private val tvShowDetailsRepository: TVShowDetailsRepository
 ) {
-    suspend operator fun invoke(id: Long): Video? {
+    suspend operator fun invoke(id: Long): String? {
         return tvShowDetailsRepository
             .getTVShowVideos(id)
             .firstOrNull { it.videoType == "Trailer" }
+            ?.videoUrl
+
     }
 }
